Updates for Solr 5.
 - remove some now-defunct attributes
 - fully qualify some filter/transformer class names

This last step to to avoid Solr bug 6188 (https://issues.apache.org/jira/browse/SOLR-6188)
wherein some strange interaction with the class loader and jars
being loaded twice causes ClassDefNotFoundErrors on core load. An
alternative way to fix it is to rename the sharedLib lib directory
and specify it explicitly in the solr.xml.